+++ tests/notebook/test_notebook_manager.py
@@ -1,6 +1,9 @@
+from textwrap import dedent
 from unittest import mock
-from unittest.mock import MagicMock
+from unittest.mock import MagicMock, PropertyMock
 
+import pytest
+from snowflake.cli.plugins.notebook.exceptions import NotebookStagePathError
 from snowflake.cli.plugins.notebook.manager import NotebookManager
 
 
@@ -21,3 +24,41 @@ def test_get_url(mock_url):
     mock_url.assert_called_once_with(
         conn_mock, f"/#/notebooks/NB_DATABASE.NB_SCHEMA.MY_NOTEBOOK"
     )
+
+
+@mock.patch("snowflake.cli.plugins.notebook.manager.make_snowsight_url")
+@mock.patch.object(NotebookManager, "_execute_queries")
+@mock.patch("snowflake.cli.plugins.notebook.manager.cli_context")
+def test_create(mock_ctx, mock_execute, mock_url):
+    type(mock_ctx.connection).warehouse = PropertyMock(return_value="MY_WH")
+    mock_url.return_value = "nb_url"
+    cn_mock = MagicMock(database="nb_db", schema="nb_schema")
+
+    with mock.patch.object(NotebookManager, "_conn", cn_mock):
+        _ = NotebookManager().create(
+            notebook_name="my_notebook",
+            notebook_file="@stage/nb file.ipynb",
+        )
+        expected_query = dedent(
+            """
+            CREATE OR REPLACE NOTEBOOK nb_db.nb_schema.my_notebook
+            FROM '@stage'
+            QUERY_WAREHOUSE = 'MY_WH'
+            MAIN_FILE = 'nb file.ipynb';
+
+            ALTER NOTEBOOK nb_db.nb_schema.my_notebook ADD LIVE VERSION FROM LAST;
+            """
+        )
+        mock_execute.assert_called_once_with(queries=expected_query)
+
+
+@pytest.mark.parametrize(
+    "stage_path",
+    (
+        pytest.param("@stage/", id="no file name"),
+        pytest.param("@stage/with/path", id="stage with path no file"),
+    ),
+)
+def test_error_parsing_stage(stage_path):
+    with pytest.raises(NotebookStagePathError):
+        NotebookManager.parse_stage_as_path(stage_path)